# contracts/openapi.yaml
openapi: "3.1.0"
info:
  title: SatQuery API
  description: |
    Earth Observation intelligence API. Upload satellite imagery, draw regions,
    ask questions about specific areas, and get answers backed by visual evidence
    and impact statistics.
  version: "1.0.0"
  contact:
    name: SatQuery Team
servers:
  - url: http://localhost:8000
    description: Local development
  - url: "{API_BASE_URL}"
    description: Configurable base URL
    variables:
      API_BASE_URL:
        default: http://localhost:8000
...
paths:
...
  /regions:
    post:
      operationId: createRegion
      tags: [regions]
      summary: Save a named region on an image
      description: |
        Creates a named region (rectangle or polygon) tied to an image.
        Geometry can be specified in GeoJSON format (if image has CRS) or
        pixel coordinates.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegionCreate"
      responses:
        "201":
          description: Region created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Region"
        "400":
          description: Invalid geometry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Image not found
  /regions/{image_id}:
    get:
      operationId: listRegions
      tags: [regions]
      summary: List all regions for an image
      parameters:
        - $ref: "#/components/parameters/ImageId"
      responses:
        "200":
          description: List of regions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Region"
  /regions/{image_id}/{region_id}:
    delete:
      operationId: deleteRegion
      tags: [regions]
      summary: Delete a region
      parameters:
        - $ref: "#/components/parameters/ImageId"
        - $ref: "#/components/parameters/RegionId"
      responses:
        "204":
          description: Region deleted
        "404":
          description: Region not found

  # ─── Query (VQA) ────────────────────────────────────────
...
  /temporal:
    post:
      operationId: submitTemporal
      tags: [temporal]
      summary: Submit a before/after image pair for change detection
      description: |
        Accepts two image IDs representing before and after states. Triggers
        background precomputation of change mask and impact statistics.
        Returns immediately with a change_id and status "pending".
        Poll GET /temporal/{change_id} for results.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemporalRequest"
      responses:
        "202":
          description: Change detection job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChangeResult"
        "400":
          description: Invalid image pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Image not found
  /temporal/{change_id}:
    get:
      operationId: getChangeResult
      tags: [temporal]
      summary: Get change detection result
      description: |
        Returns the precomputed change map, change percentage, and impact
        statistics. If still processing, returns status "processing".
      parameters:
        - $ref: "#/components/parameters/ChangeId"
      responses:
        "200":
          description: Change result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChangeResult"
        "404":
          description: Change result not found

  # ─── Optical-SAR Fusion ─────────────────────────────────
...
components:
  parameters:
    ImageId:
      name: image_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    RegionId:
      name: region_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    ChangeId:
      name: change_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, detail]
      properties:
        error:
          type: string
        detail:
          type: string

    # ─── Image ───────────────────────────────────────────
...
    RegionCreate:
      type: object
      required: [image_id, name]
      properties:
        image_id:
          type: string
          format: uuid
        name:
          type: string
          description: "Human-readable name (e.g. Region A)"
        geometry:
          $ref: "#/components/schemas/GeoJsonPolygon"
          nullable: true
          description: GeoJSON polygon (if image has CRS)
        pixel_bounds:
          $ref: "#/components/schemas/PixelBounds"
          nullable: true
          description: Pixel-coordinate bounding box
    Region:
      type: object
      required: [id, image_id, name, created_at]
      properties:
        id:
          type: string
          format: uuid
        image_id:
          type: string
          format: uuid
        name:
          type: string
        geometry:
          $ref: "#/components/schemas/GeoJsonPolygon"
          nullable: true
        pixel_bounds:
          $ref: "#/components/schemas/PixelBounds"
          nullable: true
        created_at:
          type: string
          format: date-time
    PixelBounds:
      type: object
      required: [x_min, y_min, x_max, y_max]
      properties:
        x_min:
          type: integer
        y_min:
          type: integer
        x_max:
          type: integer
        y_max:
          type: integer

    # ─── Query ───────────────────────────────────────────
...
    TemporalRequest:
      type: object
      required: [before_image_id, after_image_id]
      properties:
        before_image_id:
          type: string
          format: uuid
        after_image_id:
          type: string
          format: uuid
    ChangeResult:
      type: object
      required: [change_id, before_image_id, after_image_id, status]
      properties:
        change_id:
          type: string
          format: uuid
        before_image_id:
          type: string
          format: uuid
        after_image_id:
          type: string
          format: uuid
        change_mask_url:
          type: string
          nullable: true
          description: URL to the change mask image
        change_pct:
          type: number
          format: double
          nullable: true
          description: Percentage of area that changed
        impact_stats:
          $ref: "#/components/schemas/ImpactStats"
          nullable: true
        status:
          type: string
          enum: [pending, processing, completed, failed]
        error_message:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
    ImpactStats:
      type: object
      properties:
        affected_population:
          type: integer
          nullable: true
        affected_buildings:
          type: integer
          nullable: true
        affected_roads_km:
          type: number
          format: double
          nullable: true
        affected_area_sqkm:
          type: number
          format: double
          nullable: true
        data_sources:
          type: array
          items:
            type: string
          description: "Sources used (e.g. WorldPop, OSM)"

    # ─── Fusion ──────────────────────────────────────────
...
    GeoJsonPolygon:
      type: object
      required: [type, coordinates]
      properties:
        type:
          type: string
          enum: [Polygon]
        coordinates:
          type: array
          description: Array of linear rings
          items:
            type: array
            items:
              type: array
              items:
                type: number
              minItems: 2
              maxItems: 3